+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added a toggle-able `Always Listening` mode
   - Each command must be prefixed with an activation phrase (defaulted to `okay ghost`)
 - Added fix for `maxPower` command (didn't work in normal DIM, only in beta)
+- Added fix for search already being populated when performing a command.
 
 ### 1.1.3 - 2022-09-20 - Perk fix